run/requirements.txt
python-dotenv==1.0.1

Flask==3.0.0
gunicorn==20.1.0

SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
